requirements.txt
alabaster==1.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docutils==0.21.2
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.121.1
fastjsonschema==2.21.2
fica==0.4.1
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.10
imagesize==1.4.1
ipykernel==6.29.5
ipylab==1.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
jupytext==1.17.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
otter-grader==6.1.4
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pillow==12.0.0
platformdirs==4.3.8
playwright==1.55.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-on-whales==0.78.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
SpeechRecognition==3.14.3
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.49.3
sympy==1.14.0
tinycss2==1.4.0
torch==2.9.0
torchvision==0.24.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.3
openai>=1.0.0
sentence-transformers>=2.2.0